Reinterpret KB items and labelings as annotations
