+++ backend/annotation/models.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.db import models
 
-from problem.models import KnowledgeBase, Problem, Sentence
+from problem.models import Problem, Sentence
 
 
 class AnnotationSession(models.Model):
@@ -57,18 +57,59 @@ class ProblemAnnotation(models.Model):
     created_at = models.DateTimeField(auto_now_add=True)
 
 
-class KnowledgeBaseAnnotation(models.Model):
+class BaseAnnotation(models.Model):
     session = models.ForeignKey(
         AnnotationSession,
         on_delete=models.CASCADE,
-        related_name="kb_annotations",
+        related_name="%(class)ss",
     )
 
-    knowledge_base = models.ForeignKey(
-        KnowledgeBase,
+    problem = models.ForeignKey(
+        Problem,
         on_delete=models.CASCADE,
-        related_name="kb_annotations",
+        related_name="%(class)ss",
     )
+
+    created_at = models.DateTimeField(auto_now_add=True)
+    created_by = models.ForeignKey(
+        settings.AUTH_USER_MODEL,
+        on_delete=models.CASCADE,
+        related_name="%(class)ss_created",
+    )
+
+    removed_at = models.DateTimeField(
+        null=True,
+        blank=True,
+        help_text="When this annotation was removed from the problem.",
+    )
+    removed_by = models.ForeignKey(
+        settings.AUTH_USER_MODEL,
+        on_delete=models.CASCADE,
+        related_name="%(class)ss_removed",
+        null=True,
+        blank=True,
+        help_text="User who removed this annotation.",
+    )
+
+    notes = models.TextField(
+        blank=True,
+        help_text="Optional notes explaining why this annotation was added or removed.",
+    )
+
+    class Meta:
+        abstract = True
+
+    def is_active(self) -> bool:
+        """Check if this annotation is currently active (not removed)."""
+        return self.removed_at is None
+
+
+class KnowledgeBaseAnnotation(BaseAnnotation):
+    class Relationship(models.TextChoices):
+        EQUAL = "equal", "Equal"
+        NOT_EQUAL = "not_equal", "Not Equal"
+        SUBSET = "subset", "Subset"
+        SUPERSET = "superset", "Superset"
 
     entity1 = models.CharField(max_length=255)
 
@@ -76,11 +117,13 @@ class KnowledgeBaseAnnotation(models.Model):
 
     relationship = models.CharField(
         max_length=255,
-        choices=KnowledgeBase.Relationship.choices,
-        default=KnowledgeBase.Relationship.EQUAL,
+        choices=Relationship.choices,
+        default=Relationship.EQUAL,
     )
 
-    created_at = models.DateTimeField(auto_now_add=True)
+    def __str__(self):
+        status = "active" if self.is_active() else f"removed at {self.removed_at}"
+        return f"KB Annotation ({self.entity1} {self.relationship} {self.entity2}) on Problem {self.problem.pk} ({status})"
 
 
 class Label(models.Model):
@@ -96,67 +139,23 @@ class Label(models.Model):
         return self.text
 
 
-class Labeling(models.Model):
-    """
-    The attachment of a label to a problem.
-
-    Each time a label is attached to a problem, a new Labeling record is created.
-    When removed, the record is marked as removed (not deleted), so the history of labelings is preserved.
-    """
-
-    problem = models.ForeignKey(
-        Problem,
-        on_delete=models.CASCADE,
-        related_name="labelings",
-    )
-
+class LabelAnnotation(BaseAnnotation):
     label = models.ForeignKey(
         Label,
         on_delete=models.CASCADE,
-        related_name="labelings",
+        related_name="label_annotations",
     )
 
-    attached_at = models.DateTimeField(auto_now_add=True)
-    attached_by = models.ForeignKey(
-        settings.AUTH_USER_MODEL,
-        on_delete=models.CASCADE,
-        related_name="labelings_attached",
-    )
-
-    removed_at = models.DateTimeField(
-        null=True,
-        blank=True,
-        help_text="When this label was removed from the problem.",
-    )
-    removed_by = models.ForeignKey(
-        settings.AUTH_USER_MODEL,
-        on_delete=models.CASCADE,
-        related_name="labelings_removed",
-        null=True,
-        blank=True,
-        help_text="User who removed this label.",
-    )
-
-    notes = models.TextField(
-        blank=True,
-        help_text="Optional notes explaining why this label was added or removed.",
-    )
-
-    class Meta:
-        ordering = ["-attached_at"]
+    class Meta(BaseAnnotation.Meta):
+        ordering = ["-created_at"]
         indexes = [
             models.Index(fields=["problem", "removed_at"]),
             models.Index(fields=["label", "removed_at"]),
         ]
         permissions = [
-            ("delete_own_labeling", "Can remove own labeling from problems"),
-            ("delete_any_labeling", "Can remove any labeling from problems"),
+            ("delete_own_label_annotation", "Can remove own label annotation from problems"),
+            ("delete_any_label_annotation", "Can remove any label annotation from problems"),
         ]
-
-
-    def is_active(self) -> bool:
-        """Check if this labeling is currently active (not removed)."""
-        return self.removed_at is None
 
     def __str__(self):
         status = "active" if self.is_active() else f"removed at {self.removed_at}"
